Move mark_as_processed() to ::DBI::Stories::Processed
To avoid cyclic dependency (not supported by Python).
