+++ mediacloud/mediawords/dbi/stories/__init__.py
@@ -16,27 +16,6 @@ class McAddStoryException(Exception):
     pass
 
 
-def mark_as_processed(db: DatabaseHandler, stories_id: int) -> bool:
-    """Mark the story as processed by inserting an entry into 'processed_stories'. Return True on success."""
-
-    # FIXME upsert instead of inserting a potential duplicate
-
-    if isinstance(stories_id, bytes):
-        stories_id = decode_object_from_bytes_if_needed(stories_id)
-
-    stories_id = int(stories_id)
-
-    log.debug("Marking story ID %d as processed..." % stories_id)
-
-    try:
-        db.insert(table='processed_stories', insert_hash={'stories_id': stories_id})
-    except Exception as ex:
-        log.warning("Unable to insert story ID %d into 'processed_stories': %s" % (stories_id, str(ex),))
-        return False
-    else:
-        return True
-
-
 def is_new(db: DatabaseHandler, story: dict) -> bool:
     """Return true if this story should be considered new for the given media source.
