+++ mediacloud/mediawords/dbi/test_stories.py
@@ -1,7 +1,6 @@
 from mediawords.db import DatabaseHandler
 # noinspection PyProtectedMember
 from mediawords.dbi.stories import (
-    mark_as_processed,
     is_new,
     combine_story_title_description_text,
     get_extracted_text,
@@ -39,16 +38,6 @@ class TestStories(TestDatabaseWithSchemaTestCase):
         self.test_download['state'] = 'success'
         self.test_download['stories_id'] = self.test_story['stories_id']
         self.db().update_by_id('downloads', self.test_download['downloads_id'], self.test_download)
-
-    def test_mark_as_processed(self):
-        processed_stories = self.db().query("SELECT * FROM processed_stories").hashes()
-        assert len(processed_stories) == 0
-
-        mark_as_processed(db=self.db(), stories_id=self.test_story['stories_id'])
-
-        processed_stories = self.db().query("SELECT * FROM processed_stories").hashes()
-        assert len(processed_stories) == 1
-        assert processed_stories[0]['stories_id'] == self.test_story['stories_id']
 
     def test_is_new(self):
